refactor(rsc): consolidate module tests into dedicated inline modules
- Move serializer tests from separate file into inline #[cfg(test)] module
- Move rendering core tests from separate file into inline #[cfg(test)] module
- Move HTML rendering tests from separate file into inline #[cfg(test)] module
- Move layout rendering tests from separate file into inline #[cfg(test)] module
- Move streaming rendering tests from separate file into inline #[cfg(test)] module
- Add #[allow(clippy::disallowed_methods)] and #[allow(clippy::module_inception)] attributes to test modules
- Remove individual test file imports (pub mod tests) from module declarations
- Consolidates test organization pattern across RSC rendering and serialization modules
